L34 - gmail-calendar-agent/requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dateutil==2.8.2
openai==1.6.1
python-dotenv==1.0.0
